update circleci client and utils structure
